+++ scripts/Game/RSS/Core/SCR_RSS_SpeedCalculator.c
@@ -544,7 +544,7 @@ class SCR_RSS_SpeedCalculator
         vector pos = owner.GetOrigin();
         vector normal = SCR_TerrainHelper.GetTerrainNormal(pos, owner.GetWorld(), false, null);
         if (velocity.Length() < VELOCITY_SIGN_THRESHOLD)
-            velocity = controller.GetVelocity();
+            velocity = SCR_PlayerBaseRssApiHelper.SampleEntityVelocity(owner);
         float cosAngle = GetSlopeProjectionCos(normal, velocity);
         float effectiveSlopeDegrees = magnitude * cosAngle;
         return Math.Clamp(effectiveSlopeDegrees, -45.0, 45.0);